ASSIGNMENT 1
I had visited NALI2023 and carried out an interview session with one of the booth’s presenters.
Most of the projects that exhibited during the event are related to ICT in teaching and learning.
Through the interview session with the presenter, I understood that there are many types of OS,
and we can explore more about the OS features in the future. I also knew different types of OS
based on the machine size, single or multi processors and applications. The features of operating
systems should be explored more in the future so that we can fully explore, learn and use the
applications of operating systems by innovating ICT for teaching and learning.