ASSIGNMENT 2
I had visited a career talk on 14th November 2023. The career talk has provided clear information for us to become a better role in industry for the future. Personally, I think that all of their words are motivating for me, especially the ones from Ms. Syarifah, Ms. Izzah and Mr. Saiful, since their jobs, which are Web Developer and UI/UX designer, are closer to our programme. From the career talk I also learned that we need to keep learning all the time, keep pushing to do better that we do.