DESIGN THINKING REPORT
I had completed a report of design thinking that under a title of Software. The main application innovating by our group is UTMSmart. We had interviewed three students from UTM to get the problems facing while using this app. After that we make a prototype to solve the problems. I hope these solution will be improve in the app and reduce the problems facing. Through this task, I know more challenge about technologies of software that occur in our daily life, I hope our ideas will help people in the future. This is also an opportunity for me to increase my skills to analyst and solve problem.